Moves deployContracts to its own file
